+++ crates/common/src/raindex_client/local_db/mod.rs
@@ -1,5 +1,6 @@
 pub mod decode;
 pub mod fetch;
+pub mod helpers;
 pub mod insert;
 pub mod query;
 pub mod sync;
