+++ crates/common/src/raindex_client/local_db/sync.rs
@@ -1,3 +1,4 @@
+use super::helpers::patch_deposit_amounts_with_decimals;
 use super::token_fetch::fetch_erc20_metadata_concurrent;
 use super::{
     insert::generate_erc20_tokens_sql,
@@ -9,9 +10,7 @@ use super::{
     *,
 };
 use alloy::primitives::Address;
-use alloy::primitives::U256;
 use flate2::read::GzDecoder;
-use rain_math_float::Float;
 use reqwest::Client;
 use std::{
     collections::{HashMap, HashSet},
@@ -300,75 +299,6 @@ async fn prepare_erc20_tokens_prefix(
     })
 }
 
-fn patch_deposit_amounts_with_decimals(
-    decoded_events: serde_json::Value,
-    decimals_by_addr: &HashMap<String, u8>,
-) -> Result<serde_json::Value, LocalDbError> {
-    let events = decoded_events.as_array().ok_or_else(|| {
-        LocalDbError::CustomError("Decoded events should be an array".to_string())
-    })?;
-
-    let mut patched = Vec::with_capacity(events.len());
-    for ev in events.iter() {
-        let mut ev_clone = ev.clone();
-        let event_type = ev.get("event_type").and_then(|v| v.as_str()).unwrap_or("");
-        if event_type == "DepositV2" {
-            let obj = ev_clone.as_object_mut().ok_or_else(|| {
-                LocalDbError::CustomError("Event should be an object".to_string())
-            })?;
-            let dd = obj
-                .get_mut("decoded_data")
-                .and_then(|v| v.as_object_mut())
-                .ok_or_else(|| {
-                    LocalDbError::CustomError("Missing decoded_data in DepositV2".to_string())
-                })?;
-
-            let token = dd.get("token").and_then(|v| v.as_str()).ok_or_else(|| {
-                LocalDbError::CustomError("Missing token in DepositV2".to_string())
-            })?;
-            let token_key = token.to_ascii_lowercase();
-            let decimals = decimals_by_addr.get(&token_key).ok_or_else(|| {
-                LocalDbError::CustomError(format!(
-                    "Missing decimals for token {} required to compute deposit_amount",
-                    token
-                ))
-            })?;
-
-            let amt_hex = dd
-                .get("deposit_amount_uint256")
-                .and_then(|v| v.as_str())
-                .ok_or_else(|| {
-                    LocalDbError::CustomError(
-                        "Missing deposit_amount_uint256 in DepositV2".to_string(),
-                    )
-                })?;
-            let digits = amt_hex.strip_prefix("0x").unwrap_or(amt_hex);
-            let amount = U256::from_str_radix(digits, 16).map_err(|e| {
-                LocalDbError::CustomError(format!(
-                    "Invalid deposit_amount_uint256 '{}': {}",
-                    amt_hex, e
-                ))
-            })?;
-
-            let amount_float = Float::from_fixed_decimal(amount, *decimals).map_err(|e| {
-                LocalDbError::CustomError(format!(
-                    "Float conversion failed for deposit_amount (token {}, decimals {}): {}",
-                    token, decimals, e
-                ))
-            })?;
-
-            dd.insert(
-                "deposit_amount".to_string(),
-                serde_json::Value::String(amount_float.as_hex()),
-            );
-        }
-
-        patched.push(ev_clone);
-    }
-
-    Ok(serde_json::Value::Array(patched))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -380,6 +310,8 @@ mod tests {
             create_tables::REQUIRED_TABLES, fetch_last_synced_block::SyncStatusResponse,
             fetch_tables::TableResponse, tests::create_success_callback, LocalDbQueryError,
         };
+        use alloy::primitives::U256;
+        use rain_math_float::Float;
         use std::cell::RefCell;
         use std::rc::Rc;
         use wasm_bindgen_test::*;
